+++ src/com/newrelic/metrics/publish/Runner.java
@@ -10,12 +10,13 @@ import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.ScheduledFuture;
 import java.util.concurrent.TimeUnit;
-import java.util.logging.Level;
 
 import com.newrelic.metrics.publish.binding.Context;
 import com.newrelic.metrics.publish.binding.Request;
+import com.newrelic.metrics.publish.configuration.Config;
 import com.newrelic.metrics.publish.configuration.ConfigurationException;
 import com.newrelic.metrics.publish.configuration.SDKConfiguration;
+import com.newrelic.metrics.publish.util.Logger;
 
 /**
  * The main entry point for executing the SDK.
@@ -24,7 +25,9 @@ import com.newrelic.metrics.publish.configuration.SDKConfiguration;
  * in a loop that never returns.
  */
 public class Runner {
-    
+
+    private static Logger logger;
+
     private List<Agent> componentAgents;
     private final SDKConfiguration config;
     private int pollInterval = 60;
@@ -33,16 +36,22 @@ public class Runner {
 
     /**
      * Constructs a {@code Runner}
+     * @throws ConfigurationException if there is a configuration issue
      */
-    public Runner() {
+    public Runner() throws ConfigurationException {
         super();
         componentAgents = new LinkedList<Agent>();
 
         try {
+            Config.init();
+            Logger.init(Config.getValue("log_level", "info"),
+                        Config.getValue("log_file_path", "logs"),
+                        Config.getValue("log_file_name", "newrelic_plugin.log"),
+                        getLogLimitInKilobytes());
+            logger = Logger.getLogger(Runner.class);
             config = new SDKConfiguration();
         } catch (Exception e) {
-            Context.getLogger().log(Level.SEVERE, e.getMessage(), e);
-            throw new RuntimeException(e);
+            throw new ConfigurationException(e.getMessage());
         }
     }
 
@@ -56,17 +65,31 @@ public class Runner {
     }
 
     /**
+     * Add an {@link Agent}
+     * @param agent the {@link Agent} to be added
+     */
+    public void add(Agent agent) {
+        componentAgents.add(agent);
+    }
+
+    /**
      * Register an {@link Agent}
      * @param agent the {@link Agent} to be registered
+     * <p>
+     * This method is now deprecated and will be removed in a future release.  Use {@link Agent#add(Agent)} instead.
      */
+    @Deprecated
     public void register(Agent agent) {
         componentAgents.add(agent);
     }
 
     /**
      * Get the {@link SDKConfiguration} for the {@code Runner}
+     * <p>
+     * This class is now deprecated and will be removed in a future release. See {@link Config}.
      * @return SDKConfiguration the current {@link SDKConfiguration}
      */
+    @Deprecated
     public SDKConfiguration getConfiguration() {
         return config;
     }
@@ -78,6 +101,7 @@ public class Runner {
      */
     public void setupAndRun() throws ConfigurationException {
         setupAgents();
+        // TODO: when removing SDKConfiguration, move config validation here
         pollInterval = config.getPollInterval();
 
         ScheduledExecutorService executor = Executors.newSingleThreadScheduledExecutor();
@@ -114,18 +138,24 @@ public class Runner {
     }
 
     private void setupAgents() throws ConfigurationException {
-        Context.getLogger().fine("Setting up agents to be run");
+        logger.debug("Setting up agents to be run");
+
         createAgents();
         if(config.internalGetServiceURI() != null) {
-            Context.log(Level.INFO, "Using host: ", config.internalGetServiceURI());
+            logger.info("Metric service URI: ", config.internalGetServiceURI());
         }
 
         context = new Context();
 
-        for (Iterator<Agent> iterator = componentAgents.iterator(); iterator.hasNext();) {
+        Iterator<Agent> iterator = componentAgents.iterator();
+        while (iterator.hasNext()) {
             Agent agent = iterator.next();
-            agent.prepareToRun(context);
+
+            setupAgentContext(agent);
+
+            agent.prepareToRun();
             agent.setupMetrics();
+
             //TODO this is a really awkward place to set the license key on the request
             context.licenseKey = config.getLicenseKey();
             if(config.internalGetServiceURI() != null) {
@@ -133,6 +163,28 @@ public class Runner {
             }
             context.internalSetSSLHostVerification(config.isSSLHostVerificationEnabled());
         }
+    }
+
+    private void setupAgentContext(Agent agent) {
+        // Since this data comes from the configured agents, it needs to be initialized here.  But only set it once since
+        // all agents should share the same version.
+        if (context.agentData.version == null) {
+            context.agentData.version = agent.getVersion();
+        }
+
+        agent.getCollector().setContext(context);
+        agent.getCollector().createComponent(agent.getGUID(), agent.getAgentName());
+    }
+
+    private Integer getLogLimitInKilobytes() {
+        Integer logLimitInKiloBytes = 25600; // 25 MB
+        if (Config.getValue("log_limit_in_kbytes") instanceof String) {
+            logLimitInKiloBytes = Integer.valueOf(Config.<String>getValue("log_limit_in_kbytes"));
+        }
+        else if (Config.getValue("log_limit_in_kbytes") instanceof Number) {
+            logLimitInKiloBytes = Config.<Long>getValue("log_limit_in_kbytes").intValue();
+        }
+        return logLimitInKiloBytes;
     }
 
     /**
@@ -147,16 +199,16 @@ public class Runner {
         @Override
         public void run() {
             try {
-                Context.log(Level.FINE, "Harvest and report data");
+                logger.debug("Harvest and report data");
 
                 Request request = context.createRequest();
 
                 for (Iterator<Agent> iterator = componentAgents.iterator(); iterator.hasNext();) {
                     Agent agent = iterator.next();
                     agent.getCollector().setRequest(request);
-                    Context.log(Level.FINE, "Beginning poll cycle for agent: '", agent.getComponentHumanLabel(), "'");
+                    logger.debug("Beginning poll cycle for agent: '", agent.getAgentName(), "'");
                     agent.pollCycle();
-                    Context.log(Level.FINE, "Ending poll cycle for agent: '", agent.getComponentHumanLabel(), "'");
+                    logger.debug("Ending poll cycle for agent: '", agent.getAgentName(), "'");
                 }
 
                 request.deliver();