+++ src/com/newrelic/metrics/publish/configuration/ConfigurationException.java
@@ -32,5 +32,4 @@ public class ConfigurationException extends Exception {
         super(cause);
     }
 
-    
 }