+++ src/com/newrelic/metrics/publish/configuration/SDKConfiguration.java
@@ -1,71 +1,51 @@
 package com.newrelic.metrics.publish.configuration;
 
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.util.Properties;
-import java.util.logging.Level;
-
-import com.newrelic.metrics.publish.binding.Context;
-
-
+import com.newrelic.metrics.publish.util.Logger;
 
 /**
  * An object representation of the data read from the {@code newrelic.properties} configuration file.
+ * <p>
+ * This class is now deprecated and will be removed in a future release. See {@link Config}.
+ * @see Config
  */
+@Deprecated
 public class SDKConfiguration {
 
-    private static final String CONFIG_PROPERTY = "newrelic.platform.config.dir";
-	private static final String DEFAULT_LICENSE_KEY = "YOUR_LICENSE_KEY_HERE";
-	private static final String propertyFileName = "newrelic.properties";
-    private static final String configPath = "config";
-	
+    private static final Logger logger = Logger.getLogger(SDKConfiguration.class);
+    
+    private static final String DEFAULT_LICENSE_KEY = "YOUR_LICENSE_KEY_HERE";
+
     private String licenseKey;
     private String serviceURI;
     private boolean sslHostVerification = true;
-	
-	/**
+
+    /**
      * Constructs a {@code SDKConfiguration}
-     * @throws IOException if the {@code newrelic.properties} file does not exist
-     * @throws ConfigurationException if there is an error reading the {@code newrelic.properties} file
+     * @throws ConfigurationException if there is an error reading the {@code newrelic.json} file
      */
-    public SDKConfiguration() throws IOException, ConfigurationException {        
-        File file = getConfigurationFile();
-        
-        System.out.println("INFO: Using configuration file " + file.getAbsolutePath());
-        
-        if (!file.exists()) {
-            throw logAndThrow(file.getAbsolutePath() + " does not exist");
-        }
-        
-        Properties props = new Properties();
-        InputStream inStream = new FileInputStream(file);
-        try {
-        	props.load(inStream);
-        } finally {
-        	inStream.close();
-        }
-        
-        licenseKey = props.getProperty("licenseKey");
+    public SDKConfiguration() throws ConfigurationException {
+
+        licenseKey = Config.getValue("license_key");
         if (null == licenseKey) {
-            throw logAndThrow("licenseKey is undefined");
+            throw new ConfigurationException("license_key is undefined");
+        }
+
+        if (licenseKey.equals(DEFAULT_LICENSE_KEY)) {
+            throw new ConfigurationException("You forgot to update the New Relic license_key from '" + DEFAULT_LICENSE_KEY + "'. " +
+            		"See https://docs.newrelic.com/docs/subscriptions/license-key for more information.");
+        }
+
+        // endpoint is optional and used for debugging
+        if (Config.getValue("endpoint") != null) {
+            serviceURI = Config.getValue("endpoint");
+        }
+
+        if (Config.getValue("ssl_host_verification") != null) {
+            sslHostVerification = (Boolean) Config.getValue("ssl_host_verification");
+            logger.debug("Using SSL host verification: ", sslHostVerification);
         }
         
-        if(licenseKey.equals(DEFAULT_LICENSE_KEY)) {
-            throw logAndThrow("You forgot to update the licenseKey from '" + DEFAULT_LICENSE_KEY + "'");        	
-        }
-        
-        //host is optional and for debugging
-        if(props.containsKey("host")) {
-        	serviceURI = props.getProperty("host");
-        	Context.log(Level.INFO, "Metric service URI: ", serviceURI);
-        }
-        
-        if (props.containsKey("sslHostVerification")) {
-            sslHostVerification = Boolean.parseBoolean(props.getProperty("sslHostVerification"));
-            Context.log(Level.FINE, "Using SSL host verification: ", sslHostVerification);
-        }
+        initProxySettings();
     }
 
     /**
@@ -91,45 +71,64 @@ public class SDKConfiguration {
     public int getPollInterval() {
         return 60;
     }
-    
+
     /**
      * For debug purposes only, not for general usage by clients of the SDK
      */
-	public String internalGetServiceURI() {
-		return serviceURI;
-	}
-	
-	/**
-	 * Returns if ssl host verification is enabled. 
-	 * Adding {@code sslHostVerification} to {@code newrelic.properties}. It is {@code true} by default.
-	 * @return boolean
-	 */
-	public boolean isSSLHostVerificationEnabled() {
-	    return sslHostVerification;
-	}
-	
-	/**
-	 * Use the 'newrelic.platform.config.dir' jvm option to set the configuration directory
-	 */
-	public static String getConfigDirectory() {
-		String path = System.getProperty(CONFIG_PROPERTY);
-		if (path == null) {
-			path = configPath;
-		}
-		return path;
-	}
+    public String internalGetServiceURI() {
+        return serviceURI;
+    }
+
+    /**
+     * Returns if ssl host verification is enabled.
+     * Adding {@code sslHostVerification} to {@code newrelic.properties}. It is {@code true} by default.
+     * @return boolean
+     */
+    public boolean isSSLHostVerificationEnabled() {
+        return sslHostVerification;
+    }
+
+    /**
+     * Now deprecated. Will be removed in a future release. See {@link Config#getConfigDirectory()}
+     */
+    @Deprecated
+    public static String getConfigDirectory() {
+        return Config.getConfigDirectory();
+    }
     
-    private File getConfigurationFile() throws ConfigurationException {
-    	String path = getConfigDirectory() + File.separatorChar + propertyFileName;
-    	File file = new File(path);
-    	if (!file.exists()) {
-    		throw logAndThrow("Cannot find config file " + path);
-    	}
-    	return file;
-	}
+    private void initProxySettings() throws ConfigurationException {
+        String protocol = getServiceURIProtocol();
+        
+        if (Config.getValue("proxy_host") != null) {
+            System.setProperty(protocol + ".proxyHost", Config.<String>getValue("proxy_host"));
+            logger.info("Using proxy host: ", Config.<String>getValue("proxy_host"));
+        }
+        
+        if (Config.getValue("proxy_port") != null) {
+            if ( !(Config.getValue("proxy_port") instanceof String) ) {
+                throw new ConfigurationException("'proxy_port' must be a String");
+            } else {
+                System.setProperty(protocol + ".proxyPort", Config.<String>getValue("proxy_port"));
+                logger.info("Using proxy port: ", Config.<String>getValue("proxy_port"));
+            }
+        }
+        
+        if (Config.getValue("proxy_username") != null) {
+            System.setProperty(protocol + ".proxyUser", Config.<String>getValue("proxy_username"));
+            logger.info("Using proxy username: ", Config.<String>getValue("proxy_username"));
+        }
+        
+        if (Config.getValue("proxy_password") != null) {
+            System.setProperty(protocol + ".proxyPassword", Config.<String>getValue("proxy_password"));
+            logger.info("Using proxy password: [REDACTED]");
+        }
+    }
     
-    private ConfigurationException logAndThrow(String message) {
-        Context.log(Level.SEVERE, message);
-        return new ConfigurationException(message);
+    private String getServiceURIProtocol() {
+        String protocol = "https";
+        if (serviceURI != null) {
+            protocol = serviceURI.startsWith("https") ? "https" : "http";
+        }
+        return protocol;
     }
 }